// MainTherapi_Frontend/Therapi FrontEnd/build/api/model/accountsModel.ts
/**
 * AiTherapy
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AccountsModel { 
    name?: string | null;
    email?: string | null;
    password?: string | null;
}