// MainTherapi_Frontend/Therapi FrontEnd/build/api/model/loginResponse.ts
/**
 * AiTherapy
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LoginResponse { 
    accessToken?: string | null;
    expiresIn?: number;
    role?: string | null;
    tokenType?: string | null;
    userName?: string | null;
}